backend/requirements.txt
fastapi
uvicorn
httpx
websockets
python-jose[cryptography]
passlib[bcrypt]
python-multipart
